+++ src/data_types/mod.rs
@@ -7,6 +7,8 @@ pub mod coordinates;
 pub mod cpic_api_results;
 /// Contains definitions related to our underlying database of genes -> alleles -> variants
 pub mod database;
+/// Contains gene definition information parsed primarily from RefSeq; coordinates, strand, exons
+pub mod gene_definition;
 /// Contains mapping stats for individual alignments
 pub mod mapping;
 /// Contains utility to normalize a variant to a standard definition that is non-ambiguous